options parsing is done by base class and calls hooks to customize it
