+++ console/cocos2d_install.py
@@ -31,7 +31,6 @@ class CCPluginInstall(cocos2d.CCPlugin):
 
     @staticmethod
     def brief_description():
-        # returns a short description of this module
         return "install a project in a device"
 
     def _build_project_dir(self, project_name, display_name):
@@ -64,31 +63,8 @@ class CCPluginInstall(cocos2d.CCPlugin):
         self._run_cmd("adb uninstall \"%s\"" % package)
         self._run_cmd("adb install \"%s\"" % apk_path)
 
-
-    # will be called from the cocos2d.py script
     def run(self, argv):
         self.parse_args(argv)
         self.install_android()
 
-    def parse_args(self, argv):
-        from optparse import OptionParser
-
-        parser = OptionParser("usage: %%prog %s -s src_dir -h -v" % CCPluginInstall.plugin_name())
-        parser.add_option("-s", "--src",
-                          dest="src_dir",
-                          help="project base directory")
-        self._add_common_options(parser)
-
-        (options, args) = parser.parse_args(argv)
-
-        if options.src_dir == None:
-            raise Exception("Please set source folder with \"-s\" or \"-src\", use -h for the usage ")
-        else:
-            if os.path.exists(options.src_dir) == False:
-              raise Exception("Error: dir (%s) doesn't exist..." % (options.src_dir))
-
-
-        workingdir = os.path.dirname(inspect.getfile(inspect.currentframe()))
-
-        self.init(options, workingdir)
