+++ console/cocos2d_version.py
@@ -41,30 +41,7 @@ class CCPluginVersion(cocos2d.CCPlugin):
     		else:
     			raise cocos2d.CCPluginError("Couldn't find version info")
 
-    # will be called from the cocos2d.py script
     def run(self, argv):
-        self.parse_args(argv)
+    	self.parse_args(argv)
         self._show_versions()
-
-    def parse_args(self, argv):
-        from optparse import OptionParser
-
-        parser = OptionParser("usage: %%prog %s -s src_dir -h -v" % CCPluginVersion.plugin_name())
-        parser.add_option("-s", "--src",
-                          dest="src_dir",
-                          help="project base directory")
-        self._add_common_options(parser)
-
-        (options, args) = parser.parse_args(argv)
-
-        if options.src_dir == None:
-            raise cocos2d.CCPluginError("Please set source folder with \"-s\" or \"-src\", use -h for the usage ")
-        else:
-            if os.path.exists(options.src_dir) == False:
-              raise cocos2d.CCPluginError("Error: dir (%s) doesn't exist..." % (options.src_dir))
-
-
-        workingdir = os.path.dirname(inspect.getfile(inspect.currentframe()))
-
-        self.init(options, workingdir)
